common: rename `string_util` file, make a `term_util` ns
