+++ .vs/launch.vs.json
@@ -65,7 +65,10 @@
         "--iso_data_path",
         "${workspaceRoot}/iso_data/jak2",
         "--game",
-        "jak2"
+        "jak2",
+        "--pretty-print",
+        "--num_threads",
+        "32"
       ]
     },
     {
