+++ common/util/string_util.cpp
@@ -1,4 +1,4 @@
-#include "StringUtil.h"
+#include "string_util.h"
 
 namespace str_util {
