+++ common/util/term_util.h
@@ -1,0 +1,7 @@
+
+
+namespace term_util {
+void clear();
+int row_count();
+int col_count();
+}  // namespace term_util
